--- app/src/main/java/com/example/dungeonanddragonsapp/data/model/ItemDb.kt
@@ -1,17 +1,0 @@
-package com.example.dungeonanddragonsapp.data.model
-
-import android.net.Uri
-import com.example.dungeonanddragonsapp.presentation.ui.model.skill.Skill
-
-/**
- * @author Andrew
- */
-data class ItemDb(
-    val id: Int,
-    val index: String,
-    val name: String,
-    val fullName: String,
-    val desc: List<String>,
-    val skills: List<Skill>,
-    val url: Uri
-)
