--- app/src/main/java/com/example/dungeonanddragonsapp/presentation/ui/fragment/home/HomeViewModel.kt
@@ -1,8 +1,0 @@
-package com.example.dungeonanddragonsapp.presentation.ui.fragment.home
-
-import androidx.lifecycle.ViewModel
-
-class HomeViewModel : ViewModel() {
-
-
-}
