--- app/src/main/java/com/example/dungeonanddragonsapp/presentation/ui/util/ItemDiffUtilCallback.kt
@@ -1,17 +1,0 @@
-package com.example.dungeonanddragonsapp.presentation.ui.util
-
-import androidx.recyclerview.widget.DiffUtil
-import com.example.dungeonanddragonsapp.presentation.ui.model.item.ItemView
-
-/**
- * @author Andrew
- */
-class ItemDiffUtilCallback: DiffUtil.ItemCallback<ItemView>() {
-    override fun areItemsTheSame(oldItem: ItemView, newItem: ItemView): Boolean {
-        return oldItem == newItem
-    }
-
-    override fun areContentsTheSame(oldItem: ItemView, newItem: ItemView): Boolean {
-        return oldItem.fullName == newItem.fullName && oldItem.skills == newItem.skills
-    }
-}
