--- app/src/main/java/com/example/dungeonanddragonsapp/presentation/ui/util/Serializer.kt
@@ -1,23 +1,0 @@
-package com.example.dungeonanddragonsapp.presentation.ui.util
-
-import android.net.Uri
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
-
-/**
- * @author Andrew
- */
-object Serializer : KSerializer<Uri> {
-    override val descriptor = PrimitiveSerialDescriptor("Uri", PrimitiveKind.STRING)
-
-    override fun deserialize(decoder: Decoder): Uri {
-        return Uri.parse(decoder.decodeString())
-    }
-
-    override fun serialize(encoder: Encoder, value: Uri) {
-        encoder.encodeString(value.toString())
-    }
-}
