+++ app/src/main/java/com/example/dungeonanddragonsapp/ui/home/HomeViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.dungeonanddragonsapp.presentation.ui.fragment.notifications
+package com.example.dungeonanddragonsapp.ui.home
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
