+++ app/src/main/res/values/strings.xml
@@ -1,11 +1,6 @@
 <resources>
     <string name="app_name">DungeonAndDragonsApp</string>
-    <string name="title_home">Дом</string>
-    <string name="title_dashboard">Список</string>
-    <string name="title_notifications">Уведомления</string>
-    <string name="name_item_list">Название</string>
-    <string name="data_detail_resource_name">Название</string>
-    <string name="data_detail_resource_desc">описание</string>
-    <string name="data_detail_resource_skills">Описание скиллов</string>
-    <string name="item_list_skills_resources">Скилы</string>
+    <string name="title_home">Home</string>
+    <string name="title_dashboard">Dashboard</string>
+    <string name="title_notifications">Notifications</string>
 </resources>
